buffers: move buffers to their own named file
